Modelling Scenarios/scenario forecast folders/Epidemic Modelling v0710/requirements.txt
affine==2.3.0
alembic==1.4.2
asteval==0.9.18
attrs==19.3.0
better-exceptions==0.2.2
brotlipy==0.7.0
Cartopy==0.18.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cliff==3.1.0
cligj==0.5.0
cmaes==0.5.0
cmd2==0.8.9
colorlog==4.1.0
contextily==1.0.0
cryptography==2.9.2
cycler==0.10.0
dask==2.17.2
decorator==4.4.2
descartes==1.1.0
dload==0.6
et-xmlfile==1.0.1
eurostat==0.2.0
Fiona==1.8.13.post1
flatten-dict==0.3.0
fsspec==0.7.4
future==0.18.2
GDAL==3.0.4
geographiclib==1.50
geopandas==0.7.0
geoplot==0.4.1
geopy==1.22.0
idna==2.9
install==1.3.1
japanmap==0.0.22
jdcal==1.4.1
joblib==0.15.1
kiwisolver==1.2.0
lmfit==1.0.1
locket==0.2.0
lxml==4.5.1
Mako==1.1.3
mapclassify==2.3.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mercantile==1.1.4
munch==2.5.0
networkx==2.4
numdifftools==0.9.39
numpy==1.18.5
olefile==0.46
opencv-python==4.2.0.34
openpyxl==3.0.3
optuna==1.5.0
ordered-set==4.0.1
OWSLib==0.20.0
pandas==1.0.4
pandaSDMX==1.0.1
partd==1.1.0
pathlib2==2.3.5
patsy==0.5.1
pbr==5.4.5
Pillow==7.1.2
prettytable==0.7.2
pycparser==2.20
pydantic==1.5.1
pyepsg==0.4.0
PyLaTeX==1.3.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyproj==2.6.1.post1
pyshp==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-Levenshtein==0.12.0
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
rasterio==1.1.5
requests==2.23.0
requests-cache==0.5.2
Rtree==0.9.4
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Shapely==1.7.0
six==1.15.0
snuggs==1.4.7
SQLAlchemy==1.3.17
statsmodels==0.11.1
stevedore==2.0.0
text-unidecode==1.3
threadpoolctl==2.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
uncertainties==3.1.4
Unidecode==1.1.1
urllib3==1.25.9
wcwidth==0.2.3
xlrd==1.2.0
